GPU support partially fixed. Adjust deserialization code to properly copy the hashes into the device Merkle tree. Unusual behavior when the error tolerance is very small.

+++ include/comp_func.hpp
@@ -39,7 +39,7 @@ struct AbsoluteComp : public BaseComp<T> {
 //        uint32_t* errorBits = reinterpret_cast<uint32_t*>(&error);
 //        int errorExponent = (((*errorBits) >> 23) & 0xFF) - 127;
 //        return Kokkos::abs(b-a) <= pow(2, errorExponent);
-        return Kokkos::abs(b-a) <= tol;
+        return Kokkos::abs(static_cast<double>(b)-static_cast<double>(a)) <= tol;
     }
 };
 #endif
